Moving legacy service configuration from the MauticCoreExtension to ServicePass
